add Storage.class
