+++ src/main/java/Event.java
@@ -35,7 +35,7 @@ public class Event extends Task {
     public String toString() {
         return "[" + printType() + "]" + "[" + getIsDoneStatus() + "] "
                 + description + " " + "(from: "
-                + from + " to: " + to + ")";
+                + Ui.printTime(from) + " to: " + Ui.printTime(to) + ")";
     }
 
     @Override
